requirements/base.txt
Django==2.1
djangorestframework==3.9.0
django-bootstrap-form==3.4
